Don't free image names on pool img info, init new cache in pool img
Each volume in a pool is opened with a pool_img object created
by wrapping TSK_IMG_INFO. This pool is closed with tsk_img_close
at the end of processing, deallocating the cache and
image name which is shared with the original img_info. The next
pool opened will then segfault when trying to access either
the cache or the images array. For now, give each pool_img its
own cache and don't call tsk_img_free on each pool img so that
we don't free images until the real img_info is closed.